Removed helper as private func instead
